pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mysqlpg"
version = "0.1.0"
description = "MySQL-compatible CLI tools for PostgreSQL"
requires-python = ">=3.8"
dependencies = [
    "psycopg2-binary>=2.9",
    "prompt_toolkit>=3.0",
    "Pygments>=2.0",
]

[project.scripts]
mysqlpg = "mysqlpg.cli:main"
mysqldumppg = "mysqlpg.dumpcli:main"

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
include = ["mysqlpg*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
